refactor: change all method return value type and paramether type from int to double

+++ SoapCalculator/src/main/java/com/bazoly/calculator/Calculator.java
@@ -2,19 +2,19 @@ package com.bazoly.calculator;
 
 public class Calculator {
 	
-	public int addition(int addend1, int addend2) {
+	public double addition(double addend1, double addend2) {
 		return addend1 + addend2;
 	}
 	
-	public int subtraction(int minuend, int subtrahend) {
+	public double subtraction(double minuend, double subtrahend) {
 		return minuend - subtrahend;
 	}
 	
-	public int multiplication(int multiplier, int multiplicand) {
+	public double multiplication(double multiplier, double multiplicand) {
 		return multiplier * multiplicand;
 	}
 	
-	public int divison(int dividend, int divisor) {
+	public double divison(double dividend, double divisor) {
 		return dividend / divisor;
 	}
